refactor: refacotr search input component architecture
