--- src/components/common/SearchInput/index.tsx
@@ -1,1 +1,0 @@
-export { default } from './SearchInput'
