+++ src/components/search/SearchInput/index.tsx
@@ -1,0 +1,2 @@
+export { default } from './SearchInput'
+export type { SearchInputProps } from './SearchInput'
