+++ src/components/search/SearchInput/search-input.scss
