+++ src/styles/global/components.scss
@@ -2,8 +2,8 @@
 @import "src/components/cart/CartToggle/cart-toggle.scss";
 @import "src/components/cart/OrderSummary/order-summary.scss";
 
-// Common
-@import "src/components/common/SearchInput/search-input.scss";
+// Search
+@import "src/components/search/SearchInput/search-input.scss";
 
 // Sections
 @import "src/components/sections/BannerText/banner-text.scss";
